### 2023-01-04 #190 @whedon
[PRE REVIEW]: Ecological Forecasting and Dynamics: A graduate course on the fundamentals of time series and forecasting in ecology
**Submitting author:** @ethanwhite (<a href="http://orcid.org/0000-0001-6728-7745">Ethan P. White</a>)
**Repository:** <a href="https://github.com/weecology/forecasting-course" target ="_blank">https://github.com/weecology/forecasting-course</a>
**Version:** v2.1.0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
