### 2023-01-04 5cf871e
Add date to paper metadata
