### 2023-01-04 #190 @ethanwhite
I added the missing metadata for `date` (note that not having this [doesn't produce an error using GH Actions](https://github.com/weecology/forecasting-course/actions/runs/3839250382)) and released as `v2.1.1`.

The DOIs that are flagged are the correct DOIs:

* The first one does not resolve but is the [DOI given in the paper]()

![Screenshot from 2023-01-04 14-01-25](https://user-images.githubusercontent.com/744427/210630612-a8187712-ab47-4827-8889-2705cb211aa6.png)

* The second one resolves correctly via the url <https://doi.org/10.1890/0012-9658(2003)084[1412:TETTWU]2.0.CO;2>.
