### 2023-02-08 #190 @labarba
The DOI checker is imperfect and can give false warnings.
