### 2023-02-08 #190 @labarba
For the paper by Bada and Olusegun, you may want to add the [full URL to the PDF](https://iosrjournals.org/iosr-jrme/papers/Vol-5%20Issue-6/Version-1/I05616670.pdf) (it is an open access journal). This is locator hard to find, in fact, because strangely Google Scholar lands on the Academia.edu deposit, not the journal landing page.

Note that the full journal name is **IOSR Journal of Research & Method in Education (IOSR-JRME)** (https://iosrjournals.org/iosr-jrme.html). There is _another journal_ with the name "International Journal of Research & Method in Education," (https://www.tandfonline.com/journals/cwse20) so one has to be careful. I was a bit suspicious of this at first (is it a copy cat journal?), but the article itself is highly cited...
