### 2023-02-08 #190 @labarba
it appears that IOSR Journals are considered predatory by several sources 😬
