### 2023-02-09 #190 @ethanwhite
> it appears that IOSR Journals are considered predatory by several sources grimacing

We are citing the paper based on its content not the journal or publisher reputation. This is a well regarded paper, which, as you note, is very well cited. It's also actively referenced as a source for broad educational initiatives (e.g., https://showcase.ems.psu.edu/node/106, https://www.buffalo.edu/catt/develop/theory/constructivism.html). This high usage rate is because it is an effective introduction and review. I think that makes it the best citation here since anyone engaging with the citation in this context will likely be looking for an understandable summary and review and this is the best one I've found. That said, if citing journals by this publisher is a concern let me know and we can either supplement or replace the citation with other sources.
