### 2023-02-09 #190 @ethanwhite
> Did not mean it as a concern for citing the paper... but a possible explanation for the broken DOI deposit.

Ah, that makes sense. And yes their DOI infrastructure definitely seems sloppy at best.
