### 2023-02-09 #190 @ethanwhite
> :wave: @ethanwhite, happy to moderate and find reviewers for your submission. Feel free to provide recommendations of potential reviewers from @whedon list reviewers.

Good folks (github handles without @) to review this submission on the provided list include:
* jhollist (ecology + R)
* arokem (time-series +R)
* gvdr (ecology + R)
* jsta (ecology + R)
* fmichonneau (ecology + R)
* jwagemann (environmental science + R)
* andrewmaclachlan (environmental science + R)
