### 2023-02-09 #190 @acocac
Hi @AoifeHughes — you've expressed interest to volunteer to review for JOSE (The Journal of Open Source Education). Could you take a look at this submission and let us know if you might be able to volunteer a review for us? We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time.
Thanks! 🙏
Title: "Ecological Forecasting and Dynamics: A graduate course on the fundamentals of time series and forecasting in ecology" [View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00190/jose.00190/10.21105.jose.00190.pdf) 📄 👈
