### 2023-02-27 #198 @whedon
[REVIEW]: Ecological Forecasting and Dynamics: A graduate course on the fundamentals of time series and forecasting in ecology
**Submitting author:** @ethanwhite (<a href="http://orcid.org/0000-0001-6728-7745">Ethan P. White</a>)
**Repository:** https://github.com/weecology/forecasting-course
**Branch with paper.md** (empty if default branch):
**Version:** 2.2

- [ ] **Version:** Does the release version given match the repository release (v2.1.0)?

- [x] **Version:** Does the release version given match the repository release (v2.1.0)?

- [x] **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)
