### 2023-03-14 weecology/forecasting-course#21 @arokem
Empty "material" pages
For example, [this page](https://course.naturecast.org/lessons/intro-time-series-data/material/) looks like this:
