### 2023-03-27 #198 @arokem
Hello! I am a bit stuck on the first checklist item "Is the source for this learning module available at the [repository url](https://github.com/weecology/forecasting-course)?" This is because I am not sure what exactly I should be reviewing. In browsing through the [course website](https://course.naturecast.org/), I am finding many links to "materials" pages that don't have any content:

https://course.naturecast.org/lessons/intro-time-series-data/material/
https://course.naturecast.org/lessons/regime-shifts/material/

Or contain some links to external readings:

https://course.naturecast.org/lessons/paleo-dynamics/material/
https://course.naturecast.org/lessons/phenology/material/
https://course.naturecast.org/lessons/community-dynamics/material/

(and several others).

There are also many pages that contain links to R tutorials about the materials, which seem to be the meat of it. I think that the structure of this website may just be confusing. What a "materials" page means, for example. I think that it would be helpful to provide a "how to use this course" page that explains the structure of the site/repo and guides potential users/students a bit more. At any rate, I would like to have some clarification, before I proceed to review the materials.
