### 2023-03-27 #198 @acocac
@arokem thanks for having a look at the resource. I agree a how to use tab or section would be very beneficial to facilitate users/contributors navigation through the e-learning resource. You are alro right there are some emtpy sections. Feel free to point them in the repository issues tab. They should be removed or completed by @ethanwhite and collaborators as part of the version submitted to JOSE.

If you go to our EiC @labarba comment on this [PREVIEW submission](https://github.com/openjournals/jose-reviews/issues/170), JOSE considers videos as supplementary material that we do not review. In general, we do not review videos, or slide presentations, or other forms of OpenCourseWare (OCW), because they cannot be modified or easily reused. JOSE reviews open-source material that is reusable and allows derived works. Also, when you make an archive of the repository, the videos will not be included.
