### 2023-04-04 #198 @acocac
@ethanwhite I would appreciate if you have any comments about empty pages in the version submitted to JOSE. @arokem already opened an issue ([#21](https://github.com/weecology/forecasting-course/issues/21)) to highlight this. Thanks for your collaboration.
