### 2023-04-04 #198 @ethanwhite
From a review perspective the `Material` tabs are not the core material and so maybe need to be renamed. It is basically links to reading and background information for students. The course "source" is the `Discussion Questions`, `Instructor Notes`, and `R Tutorial` sections. While the beginning of the R tutorials is often video - if you scroll down there is a full set of written R tutorial material on each page. Hope this helps.
