### 2023-04-04 #198 @acocac
@arokem please read the comments above. Please continue the revision and checklist according to the course source suggested by @ethanwhite. Note you should follow the [guidelines for JOSE reviewers](https://joss.readthedocs.io/en/latest/review_criteria.html).
