### 2023-04-28 #198 @arokem
> Version: Does the release version given match the repository release (v2.1.0)?

I see that the repo is now at v2.1.1, but I assume that's not a problem for the review. Leaving unchecked for now until that's confirmed.
