### 2023-04-28 #198 @arokem
> Installation instructions: Is there a clearly stated list of dependencies?

The only list or instructions for installation that I could find is in the README of the repository. Unfortunately, these instructions are not for the learners/instructors, but for folks interested in building the (rather sophisticated...) website that currently serves these materials. I would suggest that these instructions should go into someplace more obscure and the README should provide a more welcoming entry point for potential learners and teachers of these materials, that would facilitate the reuse of the materials.
