### 2023-04-28 #198 @arokem
> Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?

This might be in there somewhere, but I couldn't find it. In general, I think that the structure of this repo and the associated website are really hard to understand and are not currently designed to easily facilitate reuse.
