### 2023-04-28 weecology/forecasting-course#22 @arokem
Regime shifts R tutorial is missing? 
I'd expect there to be some material here: https://course.naturecast.org/lessons/regime-shifts/r_tutorial/?

Related to https://github.com/openjournals/jose-reviews/issues/198
