### 2023-04-28 weecology/forecasting-course#23 @arokem
Book link is UFL-specific
On [this page](https://course.naturecast.org/lessons/uncertainty/material/), the link to the book chapter is https://ebookcentral.proquest.com/lib/UFL/detail.action?docID=4866481#goto_toc, which only works if you are connected through UFL. Please replace with a more general link (or at least put some kind of warning next to it?).

Related to https://github.com/openjournals/jose-reviews/issues/198
