### 2023-04-28 weecology/forecasting-course#24 @arokem
Latex rendering issues
There are some equations on this page that aren't properly rendered: https://course.naturecast.org/lessons/r-empirical-dynamic-modeling/r_tutorial/

Related to https://github.com/openjournals/jose-reviews/issues/198
