### 2023-04-28 #198 @arokem
Regarding content: I think that the content covered  and pedagogy are both really great! Clearly, the authors have a lot of experience teaching this kind of material and the coverage is really ambitious. Relatedly: I don't think that I know who the target audience for this material is. Is it designed for undergraduates or graduate students? What should they have learned as pre-requisites? Maybe this is stated somewhere, but I couldn't find it.

One suggestion for an advanced topic that you might want to cover is the use of generalized additive models for time-series modeling with mgcv. Gavin Simpson wrote a great introductory [paper](https://www.frontiersin.org/articles/10.3389/fevo.2018.00149/full) (and lots of [blog posts](https://fromthebottomoftheheap.net/blog/)) on this topic that you can use as a starting point. Admittedly, it's not a great _forecasting_ model, though, so possibly out of scope for this course.
