### 2023-04-28 #198 @arokem
The paper manuscript is really well-written, motivates the work well, and provides a clear story of how these materials came about, the design principles underlying the materials' organization, and how they might be used (with the caveat that it would be good to specify the target audience more clearly, see above).
