### 2023-04-28 #198 @arokem
OK - I am done with a review of this work. Based on my feedback, I've left the following items unchecked:

- [ ] Version: Does the release version given match the repository release (v2.1.0)?
- [ ] Installation instructions: Is there a clearly stated list of dependencies?
- [x] Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?

Arguably, the last one could be checked, but I think that the materials are not currently designed to facilitate this kind of usage. I think that the addition of some documentation on how this is all put together into a course might be sufficient to check this box as well, but just wanted to keep this in mind as you work on revisions.
