### 2023-05-15 #198 @acocac
Hi @mdietze, thanks for your interest and availability to review the submission in this REVIEW issue for JOSE.

Before starting, I appreciate to declare any relevant conflict of interest here. Please note JOSE reviewers are asked to provide constructive review with the purpose of improving the quality of the submission. We generally do not ask reviewers to “accept/reject” the submission. You should instead recommend publication after all your concerns and suggestions have been addressed. Please update your checklist available at the first comment of this REVIEW issue.

Please feel free to email back or contact in GitHub if you have any questions.
