### 2023-06-06 #198 @mdietze
* Installation instructions were clear and easy. I was able to fork the repo and launch in [netlify.com](http://netlify.com/) in only a couple minutes
* I skimmed through all of the lessons, both those that are “science” focused and those that are “code” focused and found the learning objectives to be clear, the discussion questions to be great, the materials to be appropriate, the quality to be high, and the pedagogy to be better than most (all?) of my own work.
* I was particularly happy to see the inclusion of “non-technical” sections, such as the lesson on forecast ethics and the paleoecological context
* Paper describing the course was straightforward and easy to read
* Overall, I have no requested changes that might prevent publication.
