### 2023-08-05 ad5bb4d
Remove non-existent R tutorial for regime shifts
This was always a paper discussion, but a blank R tutorial page was added by accident.

Closes #22
