+++ content/lessons/regime-shifts/r_tutorial.md
@@ -1,8 +1,0 @@
----
-title: "R Tutorial"
-weight: 3
-summary: " "
-show_date: false
-editable: true
----
-
