### 2023-08-06 30d197b
Improve lesson navigation by making all pages book
All pages within the lessons now correctly show the sidebar navigation
to allow quickly moving between different pages within a lesson.

Also fixes a couple of pages with missing show_date and editable metadata
