+++ content/lessons/LessonTemplate/r_tutorial.md
@@ -1,6 +1,7 @@
 ---
 title: "R Tutorial"
 weight: 3
+type: book
 summary: " "
 show_date: false
 editable: true
