+++ content/lessons/species-distribution-models/instructor_notes.md
@@ -1,6 +1,7 @@
 ---
 title: "Instructor Notes"
 weight: 4
+type: book
 summary: " "
 show_date: false
 editable: true
@@ -9,6 +10,7 @@ editable: true
 ---
 title: "Discussion Questions"
 weight: 2
+type: book
 summary: " "
 show_date: false
 editable: true
