+++ content/lessons/ethics/discussion_questions.md
@@ -1,6 +1,7 @@
 ---
 title: "Discussion Questions"
 weight: 2
+type: book
 summary: " "
 show_date: false
 editable: true
