+++ content/lessons/uncertainty/instructor_notes.md
@@ -1,6 +1,7 @@
 ---
 title: "Instructor Notes"
 weight: 4
+type: book
 summary: " "
 show_date: false
 editable: true
