+++ content/lessons/CourseIntroduction/instructor_notes.md
@@ -2,6 +2,7 @@
 title: "Instructor Notes"
 linkTitle: "Instructor Notes"
 weight: 4
+type: book
 summary: " "
 show_date: false
 editable: true
