+++ content/lessons/species-distribution-models/instructor_notes.md
@@ -7,15 +7,6 @@ show_date: false
 editable: true
 ---
 
----
-title: "Discussion Questions"
-weight: 2
-type: book
-summary: " "
-show_date: false
-editable: true
----
-
 ## Concepts
 
 > How (generally) do species distribution models work?
