### 2023-08-06 81f759c
Remove lesson with no content
