+++ content/lessons/regime-shifts/material.md
@@ -1,10 +1,0 @@
----
-title: "Material"
-weight: 1
-type: book
-summary: " "
-show_date: false
-editable: true
----
-
-
