### 2023-08-06 weecology/forecasting-course#21 @ethanwhite
Thanks for flagging this @arokem. All of the empty pages should now be removed via #28.
