### 2023-08-06 88ea8fa
Fix LATEX rendering for EDM tutorial
* Switch to wowchemy's built in math handling
* Handle all math terms with LATEX
* Make long equation that wasn't fitting multiline

Fixes #24
