### 2023-08-10 ade41e4
Add additional links to Dietze Ch 2
The original link only worked for UF students. This adds links to JSTOR
for the chapter and Princeton University Press to purchase the book.

Fixes #23
