### 2023-08-10 6e5b63b
Replace "Material" with more specific names
There has been some confusion to what the links to "Material" related to
on the site. This replaces that generic title with more specific titles
including Reading, Reading & Video, and Data & Software.

Addresses https://github.com/openjournals/jose-reviews/issues/198#issuecomment-1484400890
