+++ content/lessons/R-time-series-data/material.md
@@ -1,5 +1,5 @@
 ---
-title: "Material"
+title: "Data & Software"
 weight: 1
 summary: " "
 type: book
