+++ content/lessons/R-time-series-modeling/material.md
@@ -1,5 +1,5 @@
 ---
-title: "Material"
+title: "Data & Software"
 weight: 1
 type: book
 summary: " "
