+++ content/lessons/R-evaluating-forecasts/material.md
@@ -1,5 +1,5 @@
 ---
-title: "Material"
+title: "Data & Software"
 weight: 1
 type: book
 summary: Software installation requirements
