### 2023-08-10 c136b19
Add info on new Getting Started page to paper
