+++ paper.md
@@ -74,11 +74,11 @@ R tutorials follow the principles of "explicit teaching" and "scaffolding", in c
 
 ## Self-guided online learning
 
-All of the course materials are available online at <https://course.naturecast.org>. This includes links to the reading materials (papers), discussion questions, instructor notes for guiding discussions, and both written and video versions of R tutorials and associated student exercises. The R tutorials are also available as a [YouTube playlist](https://www.youtube.com/watch?v=Zr81Xn-sic4&list=PLD8eCxFKntVETvfPd-diUORGYLAL6idBv). To mirror the scaffolding approach used during in-person R tutorials, video-based lessons are typically split into a series of short videos with natural breakpoints for students to conduct exercises. To support self-guided students the video lessons review the expected output for each exercise at the beginning of the next video (i.e. immediately after the pause wherein the student is expected to complete the exercise).
+All of the course materials are available online at <https://course.naturecast.org>. This includes links to the reading materials (papers), discussion questions, instructor notes for guiding discussions, and both written and video versions of R tutorials and associated student exercises. The R tutorials are also available as a [YouTube playlist](https://www.youtube.com/watch?v=Zr81Xn-sic4&list=PLD8eCxFKntVETvfPd-diUORGYLAL6idBv). To mirror the scaffolding approach used during in-person R tutorials, video-based lessons are typically split into a series of short videos with natural breakpoints for students to conduct exercises. To support self-guided students the video lessons review the expected output for each exercise at the beginning of the next video (i.e. immediately after the pause wherein the student is expected to complete the exercise). The [Getting Started](https://course.naturecast.org/getting-started/) on the website introduces self-guided learners to approaches for engaging in the material in a self-guided context.
 
 ## Reuse and remixing
 
-Individual components of the course (e.g., individual R tutorials) and the course as a whole are designed to be reused in other classroom environments. To allow instructors to customize which material they use to suit their classroom, the course website has a modular design that allows instructors to remix the lessons into a different order or to add or remove lessons as desired. This is accomplished through  changes to the YAML markup in a single file in the website contents. The modular structure of the website thus allows other instructors to easily set up a version of the course that works for their needs and those of their students.
+Individual components of the course (e.g., individual R tutorials) and the course as a whole are designed to be reused in other classroom environments. To allow instructors to customize which material they use to suit their classroom, the course website has a modular design that allows instructors to remix the lessons into a different order or to add or remove lessons as desired. This is accomplished through  changes to the YAML markup in a single file in the website contents. The modular structure of the website thus allows other instructors to easily set up a version of the course that works for their needs and those of their students. The [Getting Started](https://course.naturecast.org/getting-started/) on the website introduces instructors to different approaches to using the material in their own courses.
 
 ## Course infrastructure
 
