### 2023-08-11 1621d78
Get the audience page to render while fixing presentation
The rending of the audience child page on the about page had been
cramped. The existing fix of using `index.md` instead of `_index.md`
for the about folder caused the page not to be rendered at all.
This fixes that while keeping a nice presentation by manually adding
line breaks.
