+++ content/about/_index.md
@@ -14,3 +14,5 @@ Course materials are openly licensed and <a href="https://github.com/weecology/f
 Anyone is welcome to use the material in their own courses and contributions are always welcome.
 
 Work on this course was supported by the National Science Foundation through grant 1929730 to S.K.M. Ernest and E.P. White, the Gordon and Betty Moore Foundation’s Data-Driven Discovery Initiative through grant GBMF4563 to E.P. White, and the University of Florida. Thanks to Claire Williams and Kelley Graff for their work scheduling and managing courses at UF and especially to Eric Hellgren for providing us the time to make the course openly available. Also, many thanks to our students over the last 7 years for their enthusiasm in learning about ecological forecasting and openness in providing feedback that has helped us improve the course with every iteration.
+<br>
+<br>
