### 2023-08-11 #198 @ethanwhite
A big thanks to @arokem & @mdietze for the helpful and generous reviews and @acocac for herding all of us cats.

We have implemented changes designed to address all of the comments both in this review issue and in the individual issues submitted by @arokem in the course repository. This includes the following changes to address @arokem's main suggestions related to providing documentation on how to adopt the lesson material:

1. We have added an audience description (including prerequisite knowledge) to both the paper (see the new Audience section) and [the website](https://course.naturecast.org/about/audience/).
2. We have added a detailed [Getting Started page](https://course.naturecast.org/getting-started/) to the website that explains how both students and teachers should start interfacing the material and linking to this in both the 'Self-guided online learning' and 'Reuse and remixing' sections of the paper.
3. We have revised the README to start with a brief [Getting Started section](https://github.com/weecology/forecasting-course#getting-started) that outlines the basics of how to adopt/adapt the material and links to the Getting Started page on the website for more details. The additional technical detail related to building the site follows below for users who want to fully leverage the infrastructure.
4. We have rename pages titled "Material" to clearly reference the specific things on those pages using the following more specific titles: "Reading", "Reading & Video", and "Data & Software".
5. Combining (3) & (4) there are now two sets of installation instructions: 1) the infrastructure installation instructions provided in the README; and 2) The "Data & Software" pages on the website which provide information on the data to download and packages to install for working through the R tutorials. These are the only two installation components to the course.
6. Regarding @arokem's suggestion of adding GAMs to the class - it's been on our [To Do list for a while](https://github.com/weecology/forecasting-course/issues/1) and we'll hopefully have time to implement it when we teach the course later this year.

Since @skmorgane & I have just completed this work we're waiting on @ha0ye to make sure he agrees with the changes, but I think given the fluid nature of JOSE review that it would be fine for @arokem to take a look at the current changes and see if they address the associated comments. Once @ha0ye and @arokem are both happy with things I'll make a new release with all of these changes (and any more that they suggest).
