+++ content/about/audience.md
@@ -5,16 +5,16 @@ summary: "Description of intended audience for this material"
 ---
 
 The long-term goal of this site is to help two large groups of people: 1) Students in a classroom at a college or university; and 2) Self-guided learners, or folks who aren’t taking a formal class and are interested in learning online on their own time.
-We are starting with the classroom students and then attempting to build on the material generated to make it maximally useful to the broader audience of folks who want to learn about ecological forecasting and dynamics.
+We are starting with the classroom students and then building on the material generated to make it maximally useful to the broader audience of folks who want to learn about ecological forecasting and dynamics.
 
 * The material is designed to be accessible to graduate students and advanced undergraduates
 * It assumes a basic ability to read and engage with the primary scientific literature, but provides guidance for engaging with each paper to help students who are learning how to do this
-* It assumes a basic understanding of R, including loading tabular data, working with variables, loading packages, and running functions. Some experience with `dplyr` and `ggplot2` is also helpful. If you need a basic introduction to R we recommend checking out the Data Carpentry lesson materials on [Data Analysis and Visualization in R for Ecologists](https://datacarpentry.org/R-ecology-lesson/).
+* It assumes a basic understanding of R, including loading tabular data, working with variables, loading packages, and running functions. Some experience with `dplyr` and `ggplot2` is also helpful. If you need a basic introduction to R, we recommend checking out the Data Carpentry lesson materials on [Data Analysis and Visualization in R for Ecologists](https://datacarpentry.org/R-ecology-lesson/).
 
 Examples of folks who we are trying to help:
 
-Maya: An advanced undergraduate in natural resources who wants to understand what ecological forecasting is and how it might be applied in conservation and management. She has used basic R in some of her other courses and has just started reading the primary scientific literature in a classroom context.
+Maya: An advanced undergraduate in natural resources who wants to understand what ecological forecasting is and how it might be applied in conservation and management. Maya has used basic R in some of their other courses and has just started reading the primary scientific literature in a classroom context.
 
 Juniper: A graduate student with a thesis related to how populations change through time, but who doesn't know anything about how to model time-series. They want to learn how to build and analyze time-series models for their thesis projects and find the idea of forecasting interesting.
 
-Jaylen: A professor who understands that ecological forecasting is becoming important for students to learn and wants to develop either a full course or a seminar on the topic. He understands the main concepts, but doesn't know what the best papers would be best for teaching and doesn't have the time to develop a bunch of R tutorials.
+Jaylen: A professor who understands that ecological forecasting is becoming important for students to learn and wants to develop either a full course or a seminar on the topic. They understand the main concepts, but don't know what the best papers would be best for teaching and don't have the time to develop a bunch of R tutorials.
