### 2023-08-14 414881e
apply consistent persona style to audience.md and paper.md come copy editing on paper.md
