+++ paper.md
@@ -33,11 +33,11 @@ bibliography: paper.bib
 
 # Summary
 
-Ecological Forecasting and Dynamics' is a semester-long course to introduce students to the fundamentals of ecological forecasting & dynamics. This course implements paper-based discussion to introduce students to concepts and ideas and R-based tutorials for hands-on application and training. The course material includes a reading list with prompting questions for discussions, teachers notes for guiding discussions, lecture notes for live coding demonstrations, and video presentations of all R tutorials. This course material can be used either as self-directed learning or as all or part of a college or university course. Individual learners have access to all of the necessary material - including discussion questions and instructor notes - on the website. The course focuses on papers with an open-access or free-to-read version where possible, though some materials still rely on access to closed-access papers. The course is structured around two sessions per week, with most weeks consisting of a one hour paper discussion session and a 1-2 hour session focused on applications in R. R tutorials use publicly available ecological datasets to provide realistic applications. Because the material is organized around content themes, instructors can modify and remix materials based on their course goals and student levels of background knowledge. These course materials have been taught for several years at the authors’ university and have also generated significant online engagement with course videos tens of thousands of times.
+'Ecological Forecasting and Dynamics' is a semester-long course to introduce students to the fundamentals of ecological forecasting & dynamics. This course implements paper-based discussion to introduce students to concepts and ideas and R-based tutorials for hands-on application and training. The course material includes a reading list with prompting questions for discussions, teachers notes for guiding discussions, lecture notes for live coding demonstrations, and video presentations of all R tutorials. This course material can be used either as self-directed learning or as all or part of a college or university course. Individual learners have access to all of the necessary material - including discussion questions and instructor notes - on the website. The course focuses on papers with an open-access or free-to-read version where possible, though some materials still rely on access to closed-access papers. The course is structured around two sessions per week, with most weeks consisting of a one hour paper discussion session and a 1-2 hour session focused on applications in R. R tutorials use publicly available ecological datasets to provide realistic applications. Because the material is organized around content themes, instructors can modify and remix materials based on their course goals and student levels of background knowledge. These course materials have been taught for several years at the authors’ university and have also generated significant online engagement with course videos tens of thousands of times.
 
 # Statement of Need
 
-Ecological forecasting is an emerging field that aims to project the current state of nature into uncertain futures. This goal of understanding and modeling nature benefits from traditional ecological approaches that assess  processes by modeling known outcomes from short-term experiments or historical data, but also involves unique tools, methods, and ways of thinking [@houlahan2017; @dietze2018; @white2019]. Many ecologists have limited exposure to all of the core concepts necessary to engage in forecasting [@brewer2003; @dietze2018] including: 1) understanding of ecological dynamics [@wolkovich2014]; the iterative cycle of model fitting, evaluation, and improvement [@dietze2018]; and 3) assessing and communicating uncertainty in forecasts [@brewer2003]. Building a community of practice around ecological forecasting requires courses that provide students with foundational conceptual knowledge relevant to ecology in conjunction with active training in methodologies and approaches [@dietze2018]. However, ecological forecasting is still a small field with few practitioners, creating a potential educational bottleneck. The ‘Ecological Forecasting and Dynamics’ course provides training in the fundamentals of ecological forecasting that will allow students to engage critically with the field and provide tools for students to deploy as they develop as forecasters. These materials can be used by instructors as modifiable core materials for their own courses or by individual students as an independent self-guided course.
[…]
