### 2023-08-15 #198 @arokem
Thanks for implementing all these revisions! I am happy to approve this paper for publication.

I spotted this awkward wording that you may choose to fix before proceeding: https://github.com/weecology/forecasting-course/pull/36

And I did still leave one check-mark unchecked:

> Does the release version given match the repository release (v2.1.0)?

Because the current release version is 2.11, but I never understood what this check was all about in the first place...
