+++ content/schedule/schedule.md
@@ -15,13 +15,13 @@ content:
     # The folders to display content from
     lessons:
       - 'Course Introduction'
-      - 'Introduction to Time Series Data'
       - 'Paleoecological Dynamics'
       - 'Working with times and dates in R'
       - 'Changes in Phenology'
       - 'Time Series Decomposition in R'
       - 'Community Dynamics'
       - 'Time Series Autocorrelation in R'
+      - 'Introduction to Time Series Data'
       - 'Time series modeling in R'
       - 'Introduction to Ecological Forecasting'
       - 'Introduction to Forecasting in R'
@@ -29,6 +29,10 @@ content:
       - 'Evaluating Forecasts in R'
       - 'Forecasting using state space models'
       - 'State space models in R'
+      - 'Introduction to Generalized Additive Models (GAMs)'
+      - 'Fitting time-series GAMS in R'
+      - 'Forecasting with GAMs in R'
+      - 'Ask us anything'
       - 'Forecasting using species distribution models'
       - 'Species Distribution Models in R'
       - 'Hurricane Forecasts'
@@ -39,30 +43,34 @@ content:
       - 'Ethics of Ecological Forecasting'
       - 'Wrap up: Can we (and what should we) forecast in ecology?'
     dates:
-      - '1'
-      - '2'
-      - '3'
-      - '4'
-      - '5'
-      - '6'
-      - '7'
-      - '8'
-      - '9'
-      - '10'
-      - '11'
-      - '12'
-      - '13'
-      - '14'
-      - '15'
-      - '16'
-      - '17'
-      - '18'
-      - '19'
-      - '20'
-      - '21'
-      - '22'
-      - '23'
-      - '24'
+      - 'August 24th'
+      - 'August 28th'
+      - 'August 30th'
+      - 'September 5th'
+      - 'September 7th'
+      - 'September 12th'
+      - 'September 14th'
+      - 'September 19th'
+      - 'September 21st'
+      - 'September 26th'
+      - 'September 28th'
+      - 'October 3rd'
+      - 'October 5th'
+      - 'October 10th'
+      - 'October 12th'
+      - 'October 17th'
+      - 'October 19th'
+      - 'October 24th'
+      - 'October 26th'
+      - 'October 31st'
+      - 'November 2nd'
+      - 'November 7th'
+      - 'November 9th'
+      - 'November 14th'
+      - 'November 16th'
+      - 'November 28th'
+      - 'November 30th'
+      - 'December 5th'
     folders:
     tag: ''
     category: ''
