### 2023-08-15 #198 @ethanwhite
I've now merged @arokem's PR (thanks @arokem!), made the equivalent change in relevant website page, and made a [new release](https://github.com/weecology/forecasting-course/releases/tag/v2.2) which is now [archived to Zenodo](https://zenodo.org/record/8250308).

@acocac - let me know if there's anything else you need on my end
