### 2023-08-18 #198 @acocac
@ethanwhite thanks for making

> Thanks for implementing all these revisions! I am happy to approve this paper for publication.
>
> I spotted this awkward wording that you may choose to fix before proceeding: [weecology/forecasting-course#36](https://github.com/weecology/forecasting-course/pull/36)
>
> And I did still leave one check-mark unchecked:
>
> > Does the release version given match the repository release (v2.1.0)?
>
> Because the current release version is 2.11, but I never understood what this check was all about in the first place...

@arokem thanks for revisiting the checklist. I have updated the version using the editorial bot. Note it doesn't update this is in the reviewers checklist, however the new PDF points to the latest version. Feel free to ignore this box and thanks for confirming your approval of the publication.
