### 2023-08-18 #198 @acocac
> I've now merged @arokem's PR (thanks @arokem!), made the equivalent change in relevant website page, and made a [new release](https://github.com/weecology/forecasting-course/releases/tag/v2.2) which is now [archived to Zenodo](https://zenodo.org/record/8250308).
>
> @acocac - let me know if there's anything else you need on my end

@ethanwhite we discussed this in the [PRE-REVIEW](https://github.com/openjournals/jose-reviews/issues/190), however I'm wondered if you can consider to change the reference from the IOSR-JRME journal. Is there any recent publication of the same author or topic alike?
