### 2023-08-21 d6684f5
Fix schedule order on website
Apparently the site is currently using weights and there's no simple way
to move that to the dates given in schedule.md.
