+++ content/schedule/schedule.md
@@ -86,8 +86,8 @@ content:
   # Useful if you wish to show the first item in the Featured widget
   offset: 0
   # Field to sort by, such as Date or Title
-  sort_by: 'Date'
-  sort_ascending: false
+  sort_by: 'Weight'
+  sort_ascending: true
 design:
   # Choose a listing view
   view: community/schedule
