+++ content/lessons/intro-to-forecasting/material.md
@@ -13,5 +13,5 @@ editable: true
 
 ## Readings
 
-* [Ecological forecasting and data assimilation in a data‐rich era](https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/09-1275.1). **Only read through the end of the "Uses of Models for Ecological Forecasting" section.** We are reading the begining this paper because it provides a good background in ecological forecasting.
+* [Ecological forecasting and data assimilation in a data‐rich era](https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/09-1275.1). **Only read through the end of the "Uses of Models for Ecological Forecasting" section.** We are reading the beginning this paper because it provides a good background in ecological forecasting.
 * [Forecasting Principles & Process Chapter 1](https://otexts.com/fpp3/intro.html). **Skip section 1.5 on case studies. Watching videos is optional**
