### 2023-08-21 9c2692a
Add stubs for new lesson material for 2023
We are adding GAMS this year but don't have the material yet. This stubs
them in for the schedule.
