+++ content/lessons/R-forecasting-gams/_index.md
@@ -1,0 +1,19 @@
+---
+title: "Forecasting with GAMs in R"
+linkTitle: ""
+type: book
+weight: 19
+summary: "Using generalized additive models for making forecasts"
+show_date: false
+editable: true
+---
+
+{{% callout note %}}
+
+**Learning Objectives:**
+* First lesson objective
+* Second lesson objective
+
+{{% /callout %}}
+
+{{< list_children >}}
