+++ content/lessons/R-species-distribution-models/_index.md
@@ -2,7 +2,7 @@
 title: "Species Distribution Models in R"
 linkTitle: "R: Species Distribution Models"
 type: book
-weight: 20
+weight: 22
 summary: An R tutorial on using fitting simple species models and making forecasts based on these models
 show_date: false
 editable: true
