+++ content/lessons/intro-to-gams/_index.md
@@ -1,0 +1,19 @@
+---
+title: "Introduction to Generalized Additive Models (GAMs)"
+linkTitle: "Intro to GAMS"
+type: book
+weight: 17
+summary: "An introduction to generalized additive models and how then can be used with time-series"
+show_date: false
+editable: true
+---
+
+{{% callout note %}}
+
+**Learning Objectives:**
+* First lesson objective
+* Second lesson objective
+
+{{% /callout %}}
+
+{{< list_children >}}
