+++ content/lessons/R-time-series-GAMS/_index.md
@@ -1,0 +1,19 @@
+---
+title: "Fitting time-series GAMS in R"
+linkTitle: ""
+type: book
+weight: 18
+summary: " "
+show_date: false
+editable: true
+---
+
+{{% callout note %}}
+
+**Learning Objectives:**
+* First lesson objective
+* Second lesson objective
+
+{{% /callout %}}
+
+{{< list_children >}}
