+++ content/lessons/species-distribution-models/material.md
@@ -9,4 +9,4 @@ editable: true
 
 ## Reading
 
-[Species Distribution Models: Ecological Explanation and Prediction Across Space and Time](http://eurobasin.dtuaqua.dk/eurobasin/documents/Training%20ISM/Elith_and_Leathwick_2009.pdf)
+[Species Distribution Models: Ecological Explanation and Prediction Across Space and Time](https://web.archive.org/web/20170809005318id_/http://sysnevo.thu.edu.tw/lib/courses_lib_file/phylogeography_file/elith_leathwick_2009.pdf)
