### 2023-08-21 0049c55
Remove link that is no longer working
See https://github.com/weecology/forecasting-course/issues/40
