+++ content/lessons/election-forecasts/material.md
@@ -7,6 +7,6 @@ show_date: false
 editable: true
 ---
 
-* Reading:[A User's Guide to FiveThirtyEight's 2016 Election Forecast](https://fivethirtyeight.com/features/a-users-guide-to-fivethirtyeights-2016-general-election-forecast/)
-* Reading:[Another model to forecast the house elections? Meet Mr. P](https://www.thecrosstab.com/2018/04/20/mrp-house/)
+* Reading: [A User's Guide to FiveThirtyEight's 2016 Election Forecast](https://fivethirtyeight.com/features/a-users-guide-to-fivethirtyeights-2016-general-election-forecast/)
+* Reading: [Another model to forecast the house elections? Meet Mr. P]()
 
