### 2023-08-21 #198 @acocac
> @acocac My understanding of how that conversation resolved was that we would still cite the paper since it is the appropriate reference. If you really want me to replace it I will, but we're also in the midst of updating the course since we're teaching with it starting Thursday and so any new release at this point to change the reference will include a significant number of additional changes. With teaching starting here there may also be some delays in getting around to tracking down an alternative. Just let me know how you'd like to proceed.

@ethanwhite, thanks for the reply and good luck with the new teaching term. EiC @labarba will decide the next steps. I trust the reference is key for the course material.
