### 2023-08-21 #198 @labarba
@acocac Please run the command `@editorialbot recommend-accept` to get everything in order for the final steps 🙏
