### 2023-08-21 #198 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
